Fix the I3 Tiling test
The test failed because one of the applications ceased to exist
on the installation media. The fix brings a different set of
applications and more to it - the applications get installed
during the test so we probably won't have to deal with similar
issue in the future. Or at least not that often.

Fixes: https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/412