post_fail_hook: also upload AVCs and a text journal dump (#410)
@kparal pointed out that it's a bit inconvenient to quickly check
the journal and any SELinux AVCs for a failed test. All the info
is in the /var/log tarball, but you have to download it, extract
it, and know how to run journalctl on it. For convenience, this
adds a text dump of the journal and one of any AVCs found with
ausearch.

Signed-off-by: Adam Williamson <awilliam@redhat.com>