Enable KDE compose tests on aarch64
This enables tests on the KDE aarch64 disk image. The test set
should be the equivalent of what we run on the Workstation disk
image - most of the same tests we run on x86_64, a few left out.
I might do a follow-up to enable the 'missing' tests on both
arches later, at least ones which don't seem to be left out for
any good reason beyond resources.

Signed-off-by: Adam Williamson <awilliam@redhat.com>