Tweak gnome-text-editor shortcuts needles
We need to replace the existing needle with one with a match area
on the 2 button, as explained in the recent commit where we
changed the test behavior. This is for Silverblue, where the old
UI is still present for now, so we need the assert_and_click on
the first shortcut page to press the '2' button for things to
work.

The removed needles all have not matched for months or years, so
let's get rid of them.

Signed-off-by: Adam Williamson <awilliam@redhat.com>