desktop_login: use validate_script_output for whoami check
The current version of this requires typing a lot of shifted
characters, which often causes typing failures at a graphical
console, especially on aarch64. We cannot pass a slower typing
speed through assert_script_run, unfortunately (I may send that
upstream as a PR later). Let's use validate_script_output and
check the output instead, because then the command is just
'whoami', which is shorter and has no shifted characters, so
less to go wrong.

Signed-off-by: Adam Williamson <awilliam@redhat.com>