static network: disable greenboot on IoT (#2396605)
In IoT F43 and F44 at present, greenboot is failing if networking
is not immediately available on boot. Well, on our tests that use
static networking, it isn't. So that's a problem. Let's have the
static networking setup test stop greenboot if it's running on
IoT, that should get around this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>