blivet software RAID: create ESP and /boot as RAID (#417)
As requested by @kparal in
https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/417 ,
this matches the modified test case by creating the EFI system
partition and /boot partition on RAID devices, as well as the
root partition. Also let's make them a bit bigger; if we leave
the value at 512 they actually get created 511MB big (dunno why)
and anaconda complains. Anyway, bigger sizes are prudent these
days.

We also need to create a BIOS boot partition on the second disk
in the BIOS test now. Apparently if /boot is RAID-ed, there has
to be a BIOS boot partition on each disk. This is a bit of an
ugly implementation, oh well.

Signed-off-by: Adam Williamson <awilliam@redhat.com>