Run desktop_keyring on KDE on updates
81d5d2f408458baf09e1afa65f0ca67670d387bd inadvertently dropped
this, then I carried that change over in the profile group
rewrite. I never intended to turn it off on KDE, though - it was
likely just a git rebase mishap or something. So let's turn it
on for KDE by using the "updates-desktops" profile group like we
do for all the others.

Signed-off-by: Adam Williamson <awilliam@redhat.com>