_ostree_build: adapt to changes in workstation-ostree-config
As of yesterday, all the *-packages.yaml files are gone. Listing
all the files that have packages: blocks in them is a pain, and
common.yaml now exists on all current branches, so let's just do
that sub in common.yaml, it should work.

Also, common.yaml no longer uses `ref: fedora/` but
`ref: ${distro}/`, so that substitution broke, which is why the
test is now failing when building the installer image (we can't
find the ostree under the ref we're expecting). Fix that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>