desktop_update_graphical: mask fwupd again, but only on GNOME
Sigh. So we stopped masking fwupd becuase it caused Discover to
blow up, but now I remember why we started doing it in the first
place: fwupd can cause another 'type' of update to appear in
GNOME Software, with an identical "Restart & Update" button,
which openQA will happily click on instead of the one for
"System Updates" which we *want* it to click on. See
https://openqa.fedoraproject.org/tests/3880370#step/desktop_update_graphical/36

So...let's go back to masking fwupd, but only on GNOME. This
should make everyone happy.

Signed-off-by: Adam Williamson <awilliam@redhat.com>