more fwupd fun: delete discover's fwupd plugin
So I'm really remembering all the fwupd/lvfs issues now. When
there's a UEFI DBX update available, we'll see it. This can cause
different issues. Right now on KDE, we're trying to install it,
but LVFS is blocking the download because it thinks we're spamming
it. So that's making tests fail.

We really need to go back to just nerfing LVFS. I think I found a
way to do it on KDE that works with current Discover: just wipe
Discover's lvfs plugin entirely. This will break if the filename
changes.

Signed-off-by: Adam Williamson <awilliam@redhat.com>