Set INSTALL_RETRY for all flavors, use on more tests, drop to 1
INSTALL_RETRY was introduced to try and retry live installs five
times to mitigate the effects of
https://bugzilla.redhat.com/show_bug.cgi?id=2329581 . That bug is
fixed, but we *do* still have some flakiness in live install
tests; we still occasionally hit the thing where install just
stalls for no obvious reason, we sometimes hit
https://bugzilla.redhat.com/show_bug.cgi?id=2402533 , and
(especially on KDE due to
https://bugzilla.redhat.com/show_bug.cgi?id=2404262 ) we sometimes
get failures caused by mistyping in language select or user
creation.

So let's keep INSTALL_RETRY around, and use it on more tests -
these are all webUI install tests that run on the live images -
but drop it to 1, 5 is probably overkill for the current known
issues.

Also, let's define it for all flavors. In the initial commit I
wrote that if it was not defined, RETRY would be set empty, but
that's not true. What actually happens is RETRY gets set to the
literal string %INSTALL_RETRY%, which is goofy. openQA still
seems to interpret it as "don't retry", which is good, but let's
avoid it.

Also move the workstation_live_iso definition into the flavors
with all the others, I don't know/remember why only that one
was in products.

Signed-off-by: Adam Williamson <awilliam@redhat.com>