Test booting ISO as USB stick
When openQA connects the iso files, it emulates the CD-ROM drive.
In the past, issues were reported that an ISO burnt to a flash drive
would not boot properly. Forcing openQA to emulate a USB flash drive
when booting and installing a system could help us catch these issue
in the future.

However, there is a small problem here. When booting from an emulated
optical drive or USB stick, openQA gives it boot priority throughout
the life of the VM. For optical install tests this isn't a problem
because anaconda automatically ejects the 'disc' at the end of the
install process, but there's no equivalent for USB, so if we just
reboot after install, we boot back to the installer.

To resolve this, after install, we go to a console, cleanly stop the
system with `systemctl halt`, then disconnect the USB stick from the
VM using a qmp command that is now exposed by os-autoinst.

Fixes: https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/34