Create test suite for Fedora Media Writer
This PR adds the test suite for Fedora Media Writer, which is
an application to download and copy the installation media files
to the a USB disk which can be then used for installation.
So far, we have been testing this application with manual testing,
this is an attempt to automate it. It tests the following:

* It chooses the Workstation media, downloads it, saves it locally
  and writes it to the USB disk. It also checks that the usual
  content exists on the USB disk.
* It opens the previously downloaded ISO and writes it to the USB disk.
  Then it check that the usual content exists on the USB disk.
* Removes the installation from the USB disk and restores it to its
  original state.
* This does not test any hotplugging mechanisms yet.

Fixes: https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/450