Update the commit reference to older app version
The remote repository for testing flatpaks has changed.
Now, it not only contains the x86_64 version of the application,
but also the aarch64 version.

The commit numbers of the specific application version have changed,
too, therefore we need to update this file and also make the difference
between the architectures.

Fixes: https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/461