Add desktop_graphics_validation test with vulkan layer check (#464)
This is a follow-up to the recent broken mesa update kerfuffle:
https://discussion.fedoraproject.org/t/173585
one outcome of that is we determined the vulkan layer validation
check actually works in a VM with llvmpipe. The command we run
here completes successfully with mesa 25.2.7-3.fc43, but crashes
with mesa 25.2.7-2.fc43. So it seems like a good idea to add this
test. I'll probably backfill a manual test case and add it to the
wiki matrix later.

I gave it a generic name in case we come up with more similar
checks we can add later. I've asked airlied for any suggestions
he might have.

Signed-off-by: Adam Williamson <awilliam@redhat.com>