Use start_with_launcher for desktop_launch_terminal on GNOME
Lately quite a lot of tests are failing here on GNOME because we
don't type terminal properly (we wind up with 'trminal' or
'teminal' usually). At first I thought my os-autoinst typing
delay change might be causing this, but it actually seems to have
started before I deployed that, so no idea what's going on. Let's
click instead of typing, to avoid the problem.

Signed-off-by: Adam Williamson <awilliam@redhat.com>