Handle known failure of vkcube --validate on Fedora 42
The new vkcube --validate test fails on Fedora 42:
https://bugzilla.redhat.com/show_bug.cgi?id=2418077
while we wait for a response to that bug, let's go with treating
this as a soft failure. We use an md5sum check to try and ensure
that the output is *exactly* what we're expecting; if there are
any unexpected messages, we'll still fail.

Signed-off-by: Adam Williamson <awilliam@redhat.com>