desktop_update_graphical: don't hit download twice
We already try to only hit download once by dropping it out of
the tags after the first time we hit it, but because we need
this check_screen to catch when *both* apply and download are
visible, we will click download twice if both apply and download
are visible on later loop iterations. So, let's guard against
that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>