Use 2M not 4M UEFI firmware images to workaround #2417734
https://bugzilla.redhat.com/show_bug.cgi?id=2417734 is about a
bug we've seen with edk2 20250812 and later where VMs quite often
seem to hang on boot, or at least boot extremely slowly so the
test times out waiting for them to reach a login screen.

We found out in testing that using the smaller-sized edk2 images
seems to avoid the problem, so let's do that until it can be
fixed properly.

Signed-off-by: Adam Williamson <awilliam@redhat.com>