Add a magic wait for the Active Directory kickstart enrol test
We get quite a few failures where the getent passwd
'SAMDOM\Administrator' check just returns nothing for some
reason. It seems like it might be a network race or something,
and this magic wait looks like it helps after a few days testing
on staging, so let's try it in prod too.

Signed-off-by: Adam Williamson <awilliam@redhat.com>