blivet custom: bump /boot and /boot/efi sizes
All the blivet custom tests are failing because anaconda is
warning about /boot being too small. Let's bump the sizes to
something more modern. 2G is the current /boot/efi default, 1G
is the floor to avoid the warning for /boot it seems like.

For software RAID we have to go 1 higher than the size we really
want as it seems like some kinda RAID overhead means we get a
partition 1 MiB smaller than we requested:
https://bugzilla.redhat.com/show_bug.cgi?id=2419063

Signed-off-by: Adam Williamson <awilliam@redhat.com>