TEMPORARY: drop services section of cockpit test
This is broken with latest Firefox, but @stransky has requested
we ignore it and allow the update to go stable as it's a security
fix. To do this we have to remove this part of the test, or else
it will fail for all subsequent updates once we allow the Firefox
updates to go stable.

Signed-off-by: Adam Williamson <awilliam@redhat.com>